Fix/app page should have router view (#28)
* 0.5.6

* #20: create hello world page in template

* #20: replace app view content and styles with router-view

* #20: configure router, include new hello-world page

* #20: configure manifest.json to generate hello-world
