+++ template/config/hello-world/index.d.ts
@@ -1,0 +1,2 @@
+export * from './hello-world';
+
